Reload ChatPage using key
